pyproject.toml
[tool.poetry]
name = "keras-training-kit"
version = "0.1.0"
description = "A template project for development, training, and evaluate DNN models with Keras"
authors = ["Takafumi Kubota <takafumi.kubota1012@gmail.com>"]
readme = "README.md"
packages = [{include = "ktk"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
tensorflow = "^2.8.0"
kaggle = "^1.5.12"
nteract-on-jupyter = "^2.1.3"
mlflow = "^1.25.1"
hydra-core = "^1.1.2"
hydra-ax-sweeper = "^1.1.5"
Pillow = "^9.1.0"
tensorflow-datasets = "^4.5.2"


[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
flake8 = "^4.0.1"
black = "^22.3.0"

[tool.black]
line-length = 100
target-version = ['py36']
include = '(\.pyi?$)'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
